+++ src/main/java/br/com/iotasoftware/gadoapp/gadoappapiv2/repository/BovineRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BovineRepository extends JpaRepository<Bovine, Long> {
-    List<Bovine> findByHerdId(Long herdId);
+public interface BovineRepository extends JpaRepository<Bovine, Integer> {
+    List<Bovine> findByHerdId(Integer herdId);
 }
